+++ code/lib/arduino-lib/HwClock.cpp
@@ -3,13 +3,13 @@
 HwClock::HwClock() {
 	char z = '0';
 
-	int day = __DATE__[5] - z + (((__DATE__[4] - z) * 10) >= 0 ? (__DATE__[4] - z) * 10 : 0);
-	int mon = 9;
-	int year = __DATE__[10] - z + (__DATE__[9] - z) * 10 + (__DATE__[8] - z) * 100 + (__DATE__[7] - z) * 1000;
+	uint8_t day = __DATE__[5] - z + (((__DATE__[4] - z) * 10) >= 0 ? (__DATE__[4] - z) * 10 : 0);
+	uint8_t mon = 9;
+	uint8_t year = __DATE__[10] - z + (__DATE__[9] - z) * 10 + (__DATE__[8] - z) * 100 + (__DATE__[7] - z) * 1000;
 
-	int hour = __TIME__[1] - z + (__TIME__[0] - z) * 10;
-	int min = __TIME__[4] - z + (__TIME__[3] - z) * 10;
-	int sec = __TIME__[7] - z + (__TIME__[6] - z) * 10;
+	uint8_t hour = __TIME__[1] - z + (__TIME__[0] - z) * 10;
+	uint8_t min = __TIME__[4] - z + (__TIME__[3] - z) * 10;
+	uint8_t sec = __TIME__[7] - z + (__TIME__[6] - z) * 10;
 
 	::setTime(hour, min, sec, day, mon, year);
 }
