+++ code/lib/arduino-lib/HwMemory.cpp
@@ -3,7 +3,7 @@
 HwMemory::HwMemory(int8_t etvNum) :
 	Memory(etvNum) {
 	if(EEPROM[0x00] != 0) {
-		int i;
+		uint8_t i;
 		for(i = 0; i < 2 * (etvNum + 1); i++)
 			EEPROM.update(i, 0);
 
@@ -30,7 +30,7 @@ int8_t HwMemory::readEtvDays(int8_t num) {
 }
 
 void HwMemory::saveStartTime(MyTime startTime) {
-	int i = 2 * _etvNum + 1 + 1;
+	uint8_t i = 2 * _etvNum + 1 + 1;
 	EEPROM.update(i++, startTime.hour);
 	EEPROM.update(i++, startTime.min);
 	EEPROM.update(i++, startTime.sec);
@@ -38,7 +38,7 @@ void HwMemory::saveStartTime(MyTime startTime) {
 
 MyTime HwMemory::readStartTime() {
 	MyTime out;
-	int i = 2 * _etvNum + 1 + 1;
+	uint8_t i = 2 * _etvNum + 1 + 1;
 
 	out.hour = EEPROM[i++];
 	out.min = EEPROM[i++];
